updated readme
doing away with jupyter notebook idea. for a couple of reasons: it would make it harder to contribute and make existing content vulnerable to changes. An easier way to contribute to the content would be strictly markdown. So, the examples folders are no more. The exercises folders are no more. And the overview will not be in jupyter notebook. 
Instead, all 3 will be individual markdown files contained within each sub-lesson's folder.